+++ changelog.md
@@ -4,13 +4,13 @@
 
 Tentative: 2023-04-04
 
-* Addition of Subjects field to metadata
-* Addition of Related Raids to metadata
 * Change public landing page title from "Raido" to "RAiD"
 * Add sign-in warning of uBlock Origin / ORCID issue
 * Add "Copy handle" button to the home page list
 * Add download for "recently minted raids" report
 * Add subjects, related Raids, related objects and alternate identifiers to metadata
+* Default the lead organisation on the mint page to the institution
+  associated with the service-point.
 
 
 # 1.0